Lighten up Organizer
It should only have public interfaces only
