+++ lib/light-service.rb
@@ -11,6 +11,7 @@ require 'light-service/organizer/with_reducer'
 require 'light-service/organizer/with_reducer_log_decorator'
 require 'light-service/organizer/with_reducer_factory'
 require 'light-service/organizer/reduce_if'
+require 'light-service/organizer/verify_call_method_exists'
 require 'light-service/action'
 require 'light-service/organizer'
 require 'light-service/orchestrator'
